♻️ Move iscn to mixin
